renaming MPIPool to TaskManager; original TaskManager moved to nbkit-batch.py
* the general tool to compute iterate over tasks with a specified task function is ``TaskManager``
* the original TaskManager code moved to nbkit-batch.py and renamed ``BatchAlgorithmDriver``
