+++ nbodykit/utils/taskmanager.py
@@ -1,324 +1,297 @@
-import logging
 import os
-import sys
-import tempfile
-from string import Formatter
+import traceback
+import logging
+import numpy
 
 from mpi4py import MPI
-from nbodykit.extensionpoints import Algorithm, algorithms
-from .parallel import MPIPool
+from nbodykit.extensionpoints import set_nbkit_comm
 
-logger = logging.getLogger('taskmanager')
- 
-def replacements_from_file(value):
+def split_ranks(N_ranks, N, include_all=False):
     """
-    Provided an existing file name, read the file into 
-    a dictionary. The keys are interpreted as the string format name, 
-    and the values are a list of values to iterate over for each job
+    Divide the ranks into chunks, attempting to have `N` ranks
+    in each chunk. This removes the master (0) rank, such 
+    that `N_ranks - 1` ranks are available to be grouped
+    
+    Parameters
+    ----------
+    N_ranks : int
+        the total number of ranks available
+    N : int
+        the desired number of ranks per worker
+    include_all : bool, optional
+        if `True`, then do not force each group to have 
+        exactly `N` ranks, instead including the remainder as well;
+        default is `False`
     """
-    if not os.path.exists(value):
-        raise RuntimeError("for `replacements_from_file`, file `%s` does not exist" %value)
-    
-    toret = {}
-    with open(value) as f:
-        code = compile(f.read(), value, 'exec')
-        exec(code, globals(), toret)
-
-    return toret
-        
-def tasks_parser(value):
-    """
-    Given a string of the format ``key tasks``, split the string and then
-    try to parse the ``tasks``, by first trying to evaluate it, and then
-    simply splitting it and interpreting the results as the tasks. 
-    
-    The general use cases are: 
-    
-    1) "box: range(2)" -> key = `box`, tasks = `[0, 1]`
-    2) "box: ['A', 'B' 'C']" -> key = `box`, tasks = `['A', 'B', 'C']`
-    """
-    import yaml
-    
-    try:
-        fields = yaml.load(value)
-        keys = list(fields.keys())
-        if len(fields) != 1:
-            raise Exception
-    except:
-        raise ValueError("specify iteration tasks via the format: ``-i key: [task1, task2]``")
-    
-    key = keys[0]
-    if isinstance(fields[key], list):
-        parsed = fields[key]
+    available = list(range(1, N_ranks)) # available ranks to do work    
+    total = len(available)
+    extra_ranks = total % N
+  
+    if include_all:
+        for i, chunk in enumerate(numpy.array_split(available, total//N)):
+            yield i, list(chunk)
     else:
-        # try to eval into a list
-        try:
-            parsed = eval(fields[key])
-            if not isinstance(parsed, list):
-                raise ValueError("result of `eval` on iteration string should be list" %(fields[key]))
-        except:
-            raise ValueError("tried but failed to `eval` iteration string `%s`" %(fields[key]))
+        for i in range(total//N):
+            yield i, available[i*N:(i+1)*N]
+    
+        if extra_ranks and extra_ranks >= N//2:
+            remove = extra_ranks % 2 # make it an even number
+            ranks = available[-extra_ranks:]
+            if remove: ranks = ranks[:-remove]
+            if len(ranks):
+                yield i+1, ranks
+            
+def enum(*sequential, **named):
+    enums = dict(zip(sequential, range(len(sequential))), **named)
+    return type('Enum', (), enums)
+    
 
-    return [key, parsed]
-
-#------------------------------------------------------------------------------
-# task manager
-#------------------------------------------------------------------------------
 class TaskManager(object):
     """
-    Task manager for running a set of `Algorithm` computations,
-    possibly in parallel using MPI
+    An MPI manager that distributes tasks over a set of MPI processes, 
+    using a specified number of independent workers to compute tasks
+    
+    Given the specified number of independent workers (which compute
+    tasks in parallel), the total number of available CPUs will be 
+    divided evenly.
+    
+    The main function is ``compute`` which tasks a list of tasks
+    and runs the task function for each item of the list
     """
-    def __init__(self, comm, 
-                       algorithm_name, 
-                       config, 
+    logger = logging.getLogger('TaskManager')
+    
+    def __init__(self, task_function, 
                        cpus_per_worker, 
-                       task_dims, 
-                       task_values, 
-                       log_level=logging.INFO, 
-                       extras={},
+                       comm=None, 
+                       debug=False, 
                        use_all_cpus=False):
         """
         Parameters
         ----------
-        comm : MPI communicator
-            the global communicator that will be split and divided
-            amongs the independent workers
-        algorithm_name : str
-            the string name of the `Algorithm` we are running
-        config : str
-            the name of the file holding the template config file, which
-            will be updated for each task that is performed
+        task_function : callable
+            the task function which takes two arguments: the task iteration
+            integer and the `task` value
         cpus_per_worker : int
             the desired number of ranks assigned to each independent
-            worker, when iterating over the tasks in parallel
-        task_dims : list
-            a list of strings specifying the names of the task dimensions -- 
-            these specify the string formatting key when updating the config
-            template file for each task value
-        task_value : list
-            a list of tuples specifying the task values which will be iterated 
-            over -- each tuple should be the length of `task_dims`
-        log_level : int, optional
-            an integer specifying the logging level to use -- default
-            is the `INFO` level
-        extras : dict, optional
-            a dictionary where the values are lists of string replacements, with
-            length equal to the total number of tasks -- if the keys are present
-            in the config file, the string formatting will update the config
-            file with the `ith` element of the list for the `ith` iteration
+            worker in the pool
+        comm : MPI communicator, optional
+            the global communicator that will be split so each worker
+            has a subset of CPUs available; default is COMM_WORLD
+        debug : bool, optional
+            if `True`, set the logging level to `DEBUG`, which prints 
+            out much more information; default is `False`
         use_all_cpus : bool, optional
-            if `True`, then do not force each worker group to have 
-            exactly `cpus_per_worker` ranks, instead including the remainder 
-            as well; default is `False`
+            if `True`, use all available CPUs, including the remainder
+            if `cpus_per_worker` is not divide the total number of CPUs
+            evenly; default is `False`
         """
-        logger.setLevel(log_level)
-        
-        self.algorithm_name  = algorithm_name
-        self.algorithm_class = getattr(algorithms, algorithm_name) 
-        self.template        = open(config, 'r').read()
+        if debug:
+            self.logger.setLevel(logging.DEBUG)
+            
+        self.function        = task_function
         self.cpus_per_worker = cpus_per_worker
-        self.task_dims       = task_dims
-        self.task_values     = task_values
-        self.extras          = extras
+        self.use_all_cpus    = use_all_cpus
         
-        self.comm      = comm
-        self.size      = comm.size
-        self.rank      = comm.rank
-        
-        # initialize the worker pool
-        kws = {'comm':self.comm, 'use_all_cpus':use_all_cpus}
-        if log_level <= logging.DEBUG: kws['debug'] = True
-        self.pool = MPIPool(self.compute_one_task, self.cpus_per_worker, **kws)
+        # the main communicator
+        self.comm      = MPI.COMM_WORLD if comm is None else comm
+        self.rank      = self.comm.rank
+        self.size      = self.comm.size
                 
-    @classmethod
-    def create(cls, comm=None, desc=None):
-        """
-        Parse the task manager and return the ``TaskManager`` instance
-        """
-        import inspect 
-        
-        if comm is None: comm = MPI.COMM_WORLD
-        args_dict = cls.parse_args(desc)
-        args_dict['comm'] = comm
-        
-        # inspect the __init__ function
-        args, varargs, varkw, defaults = inspect.getargspec(cls.__init__)
-        
-        # determine the required arguments
-        args = args[1:] # remove 'self'
-        if defaults:
-            required = args[:-len(defaults)]
-        else:
-            required = args
-            
-        # get the args, kwargs to pass to __init__
-        fargs = tuple(args_dict[p] for p in required)
-        fkwargs = {}
-        if defaults:
-            for i, p in enumerate(defaults):
-                name = args[-len(defaults)+i]
-                fkwargs[name] = args_dict.get(name, defaults[i])
-        
-        return cls(*fargs, **fkwargs)
-        
-    @classmethod
-    def parse_args(cls, desc=None):
-        """
-        Parse command-line arguments that are needed to initialize a 
-        `TaskManager` class
-        
-        Parameters
-        ----------
-        desc : str, optional
-            the description of to use for this parser
-        """
-        import argparse
-        import itertools
-        
-        # parse
-        parser = argparse.ArgumentParser(description=desc) 
-        
-        # first argument is the algorithm name
-        h = 'the name of the `Algorithm` to run in batch mode'
-        valid_algorithms = list(vars(algorithms).keys())  
-        parser.add_argument(dest='algorithm_name', choices=valid_algorithms, help=h)  
-        
-        # the number of independent workers
-        h = """the desired number of ranks assigned to each independent
-                worker, when iterating over the tasks in parallel""" 
-        parser.add_argument('cpus_per_worker', type=int, help=h)
-    
-        # now do the required named arguments
-        required_named = parser.add_argument_group('required named arguments')
-        
-        # specify the tasks along one dimension 
-        h =  """given a string of the format ``key: tasks``, split the string and then
-                try to parse the ``tasks``, by first trying to evaluate it, and then
-                simply splitting it and interpreting the results as the tasks. 
-        
-                The general use cases are: 
-        
-                1) "box: range(2)" -> key = `box`, tasks = `[0, 1]`
-                2) "box: [A, B, C]" -> key = `box`, tasks = `['A', 'B', 'C']`
-                
-                If multiple options passed with `-i` flag, then the total tasks to 
-                perform will be the product of the tasks lists passed"""
-        required_named.add_argument('-i', dest='tasks', action='append', 
-                type=tasks_parser, required=True, help=h)
-    
-        # the template config file
-        h = """the name of the template config file (using YAML synatx) that 
-                provides the `Algorithm` parameters; the file should use 
-                ``string.format`` syntax to indicate which variables will be 
-                updated for each task, i.e., an input file could be specified 
-                as 'input/DataSource_box{box}.dat', if `box` were one of the task 
-                dimensions"""
-        required_named.add_argument('-c', '--config', type=str, help=h)
-    
-        # read any extra string replacements from file
-        h = """file providing extra string replaces, with lines of the form 
-                 `tag = ['tag1', 'tag2']`; if the keys match keywords in the 
-                 template param file, the file with be updated with
-                 the `ith` value for the `ith` task"""
-        parser.add_argument('--extras', dest='extras', default={}, type=replacements_from_file, help=h)
-    
-        h = "set the logging output to debug, with lots more info printed"
-        parser.add_argument('--debug', help=h, action="store_const", dest="log_level", 
-                            const=logging.DEBUG, default=logging.INFO)
-                            
-        h = "if `True`, include all available cpus in the worker pool"
-        parser.add_argument('--use_all_cpus', action='store_true', help=h)
-                                
-        args = parser.parse_args()
-        
-        # format the tasks, taking the product of multiple task lists
-        keys = []; values = []
-        for [key, tasks] in args.tasks:
-            keys.append(key)
-            values.append(tasks)
+        # need at least one
+        if self.size == 1:
+            raise ValueError("need at least two processes to use a TaskManager")
 
-        # take the product
-        if len(keys) > 1:
-            values = list(itertools.product(*values))
-        else:
-            values = values[0]
-            
-        # save
-        args.task_dims = keys
-        args.task_values = values
+        # make the sub-communicator
+        self._create_subcomm()
+        self.status = MPI.Status()
         
-        return vars(args)
-            
-    def compute(self):
+        # communication tags
+        self.tags = enum('READY', 'DONE', 'EXIT', 'START')
+
+    def _create_subcomm(self):
         """
-        Compute all of the tasks
-        """             
-        # compute the work
-        results = self.pool.compute(self.task_values)
-        return results
-            
-    def compute_one_task(self, itask, task):
+        Internal function to create the sub-communicator that will be 
+        used by each independent worker when running the task function
+        
+        Notes
+        -----
+        This function updates the global ``nbodykit`` communicator
+        via ``set_nbkit_comm`` so all plugins will be initialized 
+        with this sub-communicator
         """
-        Run the algorithm once, using the parameters specified by `task`,
-        which is the `itask` iteration
+        self.subcomm = None
+        chain_ranks = []
+        color = 0
+        total_ranks = 0
+        
+        # split the ranks
+        for i, ranks in split_ranks(self.size, self.cpus_per_worker, include_all=self.use_all_cpus):
+            chain_ranks.append(ranks[0])
+            if self.rank in ranks: color = i+1
+            total_ranks += len(ranks)
+        
+        self.workers = i+1 # store the total number of workers
+        leftover = (self.size - 1) - total_ranks
+        if leftover and self.rank == 0:
+            args = (self.cpus_per_worker, self.size-1, leftover)
+            self.logger.warning("with `cpus_per_worker` = %d and %d available ranks, %d ranks will do no work" %args)
+            
+        # crash if we only have one process or one worker
+        if self.size <= self.workers:
+            args = (self.size, self.workers+1, self.workers)
+            raise ValueError("only have %d ranks; need at least %d to use the desired %d workers" %args)
+            
+        # ranks that will do work have a nonzero color now
+        self._valid_worker = color > 0
+        
+        # split the comm between the workers
+        self.subcomm = self.comm.Split(color, 0)
+        
+        # set the global extension point comm
+        set_nbkit_comm(self.subcomm)
+        
+    def is_master(self):
+        """
+        Is the current process the master?
+        """
+        return self.rank == 0
+
+    def is_worker(self):
+        """
+        Is the current process a valid worker (and thus should wait for
+        instructions from the master)
+        """
+        return self._valid_worker
+        
+    def wait(self):
+        """
+        If this isn't the master process, wait for instructions.
+        """
+        if self.is_master():
+            raise RuntimeError("master node told to await jobs")
+
+        # logging info
+        if self.subcomm.rank == 0:
+            args = (self.rank, MPI.Get_processor_name(), self.subcomm.size)
+            self.logger.info("worker master rank is %d on %s with %d processes available" %args)
+
+        # continously loop and wait for instructions
+        while True:
+            args = None
+            tag = -1
+
+            # have the master rank of the subcomm ask for task and then broadcast
+            if self.subcomm.rank == 0:
+                self.comm.send(None, dest=0, tag=self.tags.READY)
+                args = self.comm.recv(source=0, tag=MPI.ANY_TAG, status=self.status)
+                tag = self.status.Get_tag()
+            
+            # bcast to everyone in the worker subcomm
+            args  = self.subcomm.bcast(args)
+            tag   = self.subcomm.bcast(tag)
+
+            # do the work here
+            if tag == self.tags.START:
+                result = self.function(*args)
+                self.subcomm.Barrier() # wait for everyone in subcomm
+                if self.subcomm.rank == 0:
+                    self.comm.send([args[0], result], dest=0, tag=self.tags.DONE) # done this task
+            elif tag == self.tags.EXIT:
+                break
+
+        # wait for everyone in subcomm and exit
+        self.subcomm.Barrier()
+        if self.subcomm.rank == 0:
+            self.comm.send(None, dest=0, tag=self.tags.EXIT) # exiting
+            
+        # debug logging
+        self.logger.debug("rank %d process is done waiting" %self.rank)
+
+    def compute(self, tasks):
+        """
+        Compute a series of tasks. For each task, the function takes
+        the iteration number, followed by the `task` value as the 
+        arguments
+        
+        Parameter
+        ---------
+        tasks : list
+            list of `task` items that will be pickled and set
+            to each worker when computing the `ith` task
+
+        Returns
+        -------
+        results : list
+            a list of the return values of the task function for 
+            each task
+        """
+        ntasks = len(tasks)
+        results = []
+
+        # catch errors
+        try:
     
-        Parameters
-        ----------
-        itask : int
-            the integer index of this task
-        task : tuple
-            a tuple of values representing this task value
-        """
-        # if you are the pool's root, write out the temporary parameter file
-        this_config = None
-        if self.pool.subcomm.rank == 0:
-            
-            # extract the keywords that we need to format from template file
-            formatter = Formatter()
-            kwargs = [kw for _, kw, _, _ in formatter.parse(self.template) if kw]
+            # master distributes the tasks and tracks closed workers
+            if self.is_master():
                 
-            # initialize a temporary file
-            with tempfile.NamedTemporaryFile(delete=False) as ff:
+                # initialize
+                task_index = 0
+                closed_workers = 0
                 
-                this_config = ff.name
-                logger.debug("creating temporary file: %s" %this_config)
+                # logging info
+                args = (self.workers, ntasks)
+                self.logger.info("master starting with %d worker(s) with %d total tasks" %args)
                 
-                # key/values for this task 
-                if len(self.task_dims) == 1:
-                    possible_kwargs = {self.task_dims[0] : task}
-                else:
-                    possible_kwargs = dict(zip(self.task_dims, task))
-                    
-                # any extra key/value pairs for this tasks
-                if self.extras is not None:
-                    for k in self.extras:
-                        possible_kwargs[k] = self.extras[k][itask]
+                # loop until all workers have finished with no more tasks
+                while closed_workers < self.workers:
+                
+                    data = self.comm.recv(source=MPI.ANY_SOURCE, tag=MPI.ANY_TAG, status=self.status)
+                    source = self.status.Get_source()
+                    tag = self.status.Get_tag()
+    
+                    # worker is ready, so send it a task
+                    if tag == self.tags.READY:
                         
-                # do the string formatting if the key is present in template
-                valid = {k:possible_kwargs[k] for k in possible_kwargs if k in kwargs}
-                ff.write(self.template.format(**valid).encode())
-        
-        # bcast the file name to all in the worker pool
-        this_config = self.pool.subcomm.bcast(this_config, root=0)
-
-        # configuration file passed via -c
-        params, extra = Algorithm.parse_known_yaml(self.algorithm_name, open(this_config, 'r').read())
-        
-        # output is required
-        output = getattr(extra, 'output', None)
-        if output is None:
-            raise ValueError("argument `output` is required in config file")
+                        # still more tasks to compute
+                        if task_index < ntasks:
+                            this_task = [task_index, tasks[task_index]]
+                            self.comm.send(this_task, dest=source, tag=self.tags.START)
+                            self.logger.info("sending task `%s` to worker %d" %(str(tasks[task_index]), source))
+                            task_index += 1
+                        # all tasks sent -- tell worker to exit
+                        else:
+                            self.comm.send(None, dest=source, tag=self.tags.EXIT)
+                    # store the results from finished tasks
+                    elif tag == self.tags.DONE:
+                        results.append(data)
+                        self.logger.debug("received result from worker %d" %source)
+                    # track workers that exited
+                    elif tag == self.tags.EXIT:
+                        closed_workers += 1
+                        self.logger.debug("worker %d has exited, closed workers = %d" %(source, closed_workers))
             
-        # initialize the algorithm and run
-        alg = self.algorithm_class(**vars(params))
-        result = alg.run()
-        alg.save(output, result)
-
-        # remove temporary files
-        if self.pool.subcomm.rank == 0:
-            if os.path.exists(this_config): 
-                logger.debug("removing temporary file: %s" %this_config)
-                os.remove(this_config)
-                
-        return 0
+            # workers will wait for instructions           
+            elif self.is_worker():
+                self.wait()
+                    
+        except Exception as e:
+            self.logger.error("an exception has occurred on one of the ranks...all ranks exiting")
+            self.logger.error(traceback.format_exc())
+            
+            # bit of hack that forces mpi4py to exit all ranks
+            # see https://groups.google.com/forum/embed/#!topic/mpi4py/RovYzJ8qkbc
+            os._exit(1)  
+            
+        finally:
+            # wait and exit
+            self.logger.debug("rank %d process finished" %self.rank)
+            self.comm.Barrier()
+            
+            if self.is_master():
+                self.logger.info("master is finished; terminating")
+                if self.subcomm is not None:
+                    self.subcomm.Free()
+            
+        # return the results in sorted order
+        results = self.comm.bcast(results)
+        return [r[1] for r in sorted(results, key=lambda r: r[0])]
